move models into database folder
