+++ backend/src/db/models/Badge.ts
