+++ backend/src/db/models/Image.ts
