+++ backend/src/db/neo4j.ts
@@ -6,7 +6,7 @@ import neo4j, { Driver } from 'neo4j-driver'
 import Neode from 'neode'
 
 import CONFIG from '@config/index'
-import models from '@models/index'
+import models from '@db/models/index'
 
 let driver: Driver
 const defaultOptions = {
