+++ backend/src/middleware/permissionsMiddleware.ts
@@ -7,8 +7,8 @@
 import { rule, shield, deny, allow, or, and } from 'graphql-shield'
 
 import CONFIG from '@config/index'
+import SocialMedia from '@db/models/SocialMedia'
 import { getNeode } from '@db/neo4j'
-import SocialMedia from '@models/SocialMedia'
 import { validateInviteCode } from '@schema/resolvers/transactions/inviteCodes'
 
 const debug = !!CONFIG.DEBUG
