+++ backend/tsconfig.json
@@ -37,7 +37,6 @@
       "@helpers/*": ["./src/helpers/*"],
       "@jwt/*": ["./src/jwt/*"],
       "@middleware/*": ["./src/middleware/*"],
-      "@models/*": ["./src/models/*"],
       "@schema/*": ["./src/schema/*"],
       "@src/*": ["./src/*"],
       "@root/*": ["./*"]
